+++ src/api/client.js
@@ -5,8 +5,8 @@ import { useAuthStore } from '@/stores/auth.js'
 const client = axios.create({
     baseURL: 'http://bookrating.test/api',
     headers: {
-        'Content-Type': 'multipart/form-data'
-    }
+        'Content-Type': 'multipart/form-data',
+    },
 })
 
 const authStore = useAuthStore()
