+++ src/assets/css/main.css
@@ -1,7 +1,6 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-
 .font-bold {
     color: #265073;
 }
