+++ src/assets/css/output.css
@@ -652,6 +652,10 @@ video {
   top: 100%;
 }
 
+.top-3 {
+  top: 0.75rem;
+}
+
 .z-10 {
   z-index: 10;
 }
@@ -668,8 +672,8 @@ video {
   grid-row: span 2 / span 2;
 }
 
-.m-3 {
-  margin: 0.75rem;
+.m-2 {
+  margin: 0.5rem;
 }
 
 .m-4 {
@@ -678,10 +682,6 @@ video {
 
 .m-5 {
   margin: 1.25rem;
-}
-
-.m-2 {
-  margin: 0.5rem;
 }
 
 .mx-5 {
@@ -702,6 +702,11 @@ video {
 .my-8 {
   margin-top: 2rem;
   margin-bottom: 2rem;
+}
+
+.mx-2 {
+  margin-left: 0.5rem;
+  margin-right: 0.5rem;
 }
 
 .mb-1 {
@@ -768,8 +773,8 @@ video {
   margin-top: 2rem;
 }
 
-.mt-1 {
-  margin-top: 0.25rem;
+.mb-0 {
+  margin-bottom: 0px;
 }
 
 .block {
@@ -790,10 +795,6 @@ video {
 
 .grid {
   display: grid;
-}
-
-.contents {
-  display: contents;
 }
 
 .hidden {
@@ -855,8 +856,8 @@ video {
   height: 100vh;
 }
 
-.h-24 {
-  height: 6rem;
+.h-8 {
+  height: 2rem;
 }
 
 .min-h-screen {
@@ -907,11 +908,6 @@ video {
   width: 24rem;
 }
 
-.w-fit {
-  width: -moz-fit-content;
-  width: fit-content;
-}
-
 .w-full {
   width: 100%;
 }
@@ -920,12 +916,8 @@ video {
   width: 100vw;
 }
 
-.w-1\/3 {
-  width: 33.333333%;
-}
-
-.w-24 {
-  width: 6rem;
+.w-8 {
+  width: 2rem;
 }
 
 .max-w-6xl {
@@ -966,10 +958,6 @@ video {
 
 .items-center {
   align-items: center;
-}
-
-.justify-end {
-  justify-content: flex-end;
 }
 
 .justify-center {
@@ -1046,20 +1034,12 @@ video {
   border-radius: 0.75rem;
 }
 
-.rounded-md {
-  border-radius: 0.375rem;
-}
-
-.border-4 {
-  border-width: 4px;
-}
-
 .border-2 {
   border-width: 2px;
 }
 
-.border {
-  border-width: 1px;
+.border-4 {
+  border-width: 4px;
 }
 
 .border-s-4 {
@@ -1146,6 +1126,14 @@ video {
 
 .p-8 {
   padding: 2rem;
+}
+
+.p-10 {
+  padding: 2.5rem;
+}
+
+.p-5 {
+  padding: 1.25rem;
 }
 
 .px-4 {
@@ -1336,46 +1324,6 @@ video {
 
 .duration-300 {
   transition-duration: 300ms;
-}
-
-@font-face {
-  font-family: 'Poppins';
-
-  src: url('/src/assets/fonts/Poppins/Poppins-Regular.ttf') format('truetype');
-
-  font-weight: 400;
-
-  font-style: normal;
-}
-
-@font-face {
-  font-family: 'Poppins';
-
-  src: url('/src/assets/fonts/Poppins/Poppins-SemiBold.ttf') format('truetype');
-
-  font-weight: 600;
-
-  font-style: normal;
-}
-
-@font-face {
-  font-family: 'Poppins';
-
-  src: url('/src/assets/fonts/Poppins/Poppins-Bold.ttf') format('truetype');
-
-  font-weight: 700;
-
-  font-style: normal;
-}
-
-@font-face {
-  font-family: 'Poppins';
-
-  src: url('/src/assets/fonts/Poppins/Poppins-Italic.ttf') format('truetype');
-
-  font-weight: 400;
-
-  font-style: italic;
 }
 
 .font-bold {
