+++ src/views/Login.vue
@@ -1,5 +1,5 @@
 <script setup>
-import Svg from '@/components/common/Image.vue'
+import Image from '@/components/common/Image.vue'
 import LoginForm from '@/components/Login/LoginForm.vue'
 </script>
 <template>
@@ -13,10 +13,10 @@ import LoginForm from '@/components/Login/LoginForm.vue'
             </p>
         </div>
         <div class="absolute bottom-0 left-0">
-            <Svg name="login4" ext="svg" w="350px" h="350px" />
+            <Image name="login4" ext="svg" w="350px" h="350px" />
         </div>
         <div class="absolute bottom-4 right-4">
-            <Svg name="login2" ext="svg" w="450px" h="450px" />
+            <Image name="login2" ext="svg" w="450px" h="450px" />
         </div>
     </div>
 </template>
